requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.3.9
chardet==3.0.4
Django==1.11
django-celery==3.2.2
django-haystack==2.8.1
django-redis==4.4.0
django-redis-sessions==0.6.1
django-tinymce==2.6.0
idna==2.8
itsdangerous==1.1.0
jieba==0.39
kombu==3.0.37
mutagen==1.42.0
Pillow==3.4.1
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.0
pytz==2018.9
redis==2.10.6
requests==2.21.0
urllib3==1.24.1
vine==1.2.0
Whoosh==2.7.4
wincertstore==0.2
